requirements.txt
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.6
chardet==3.0.4
click==7.1.2
Flask==2.0.1
flask-ngrok==0.0.25
gevent==21.8.0
google-api-core==1.22.2
google-auth==1.21.1
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
greenlet==1.1.1
gunicorn==20.1.0
httplib2==0.18.1
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.21.2
Pillow==8.3.1
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pytz==2020.1
requests==2.24.0
rsa==4.6
six==1.15.0
uritemplate==3.0.1
urllib3==1.25.10
Werkzeug==2.0.1
zope.event==4.5.0
zope.interface==5.4.0
